+++ assets/license-page.stories-DupQJNop.js
@@ -1,0 +1,57 @@
+import{r as p,j as m}from"./iframe-DKr_oYML.js";import{L as r}from"./LicensePage-CkMMVSgA.js";import"./index-DPjhLrE2.js";import"./FilledIcon-B2fkJSSc.js";import"./BackIcon-BM77X_6n.js";import"./type-CgT0EiC8.js";import"./index-CN4RR1pw.js";import"./Heading-ZSzZQJid.js";import"./index-BdRzX_qh.js";import"./InfoIcon-C46TqUho.js";import"./Text-ClZBMQ7l.js";import"./CheckmarkIcon-C-jEfIIV.js";import"./WelcomeIcon-bM0LKDFe.js";import"./index-BgDyFtBm.js";import"./CloseIcon-CLW0AMie.js";import"./index-Bjo1Nh62.js";import"./StopIcon-CWxrEcvX.js";import"./index-CofpwxeZ.js";import"./Label--_eoK5Eh.js";import"./index-DvAn6bT-.js";import"./index-Dq7h7Pqt.js";import"./TextLink-DDpIuKCx.js";import"./UpdatesIcon-BT6dDZDt.js";import"./OpenIcon-Cgd0wLCq.js";const Q={title:"PAC Universal Plugin UI/Pages/License Page",component:r,parameters:{pageLayout:"page"}},a=()=>{const[e,t]=p.useState({key:"",licence_status:"",licence_expire:"",last_check:null}),s={isFromDiviMarket:!1,activateLicenseApi:"https://dummyjson.com/c/0a30-b68a-42b3-88c0",forceLicenseCheckApi:"https://dummyjson.com/c/f33c-6d65-412e-8244",deactivateLicenseApi:"https://dummyjson.com/c/03c3-39d0-45ae-9613",licenseData:e,setLicenseData:c=>t(c),findLicenseKeyLink:"https://www.peeayecreative.com/account/licence-keys/divi-assistant/",docsLink:"https://www.peeayecreative.com/docs/divi-assistant/troubleshooting-faqs/"};return m.jsx(r,{...s})},i=()=>{const[e,t]=p.useState({key:"",licence_status:"",licence_expire:"",last_check:null}),s={isFromDiviMarket:!1,activateLicenseApi:"https://dummyjson.com/c/21ea-fa9c-4fde-b725",forceLicenseCheckApi:"https://dummyjson.com/c/f33c-6d65-412e-8244",deactivateLicenseApi:"https://dummyjson.com/c/03c3-39d0-45ae-9613",licenseData:e,setLicenseData:c=>t(c),findLicenseKeyLink:" https://www.peeayecreative.com/account/licence-keys/divi-assistant/",docsLink:"https://www.peeayecreative.com/docs/divi-assistant/troubleshooting-faqs/"};return m.jsx(r,{...s})},n=()=>{const[e,t]=p.useState({key:"PADTMAU-c26a58cc-e3b-2d756-b497591e",licence_status:"active",licence_expire:"2023-12-31",last_check:Math.floor(Date.now()/1e3)}),s={isFromDiviMarket:!1,activateLicenseApi:"https://dummyjson.com/c/0a30-b68a-42b3-88c0",forceLicenseCheckApi:"https://dummyjson.com/c/f33c-6d65-412e-8244-ss",deactivateLicenseApi:"https://dummyjson.com/c/03c3-39d0-45ae-9613",licenseData:e,setLicenseData:c=>t(c),findLicenseKeyLink:" https://www.peeayecreative.com/account/licence-keys/divi-assistant/",docsLink:"https://www.peeayecreative.com/docs/divi-assistant/troubleshooting-faqs/"};return m.jsx(r,{...s})},o=()=>m.jsx(r,{isFromDiviMarket:!0});a.__docgenInfo={description:"",methods:[],displayName:"NotActivated"};i.__docgenInfo={description:"",methods:[],displayName:"InvalidKey"};n.__docgenInfo={description:"",methods:[],displayName:"Activated"};o.__docgenInfo={description:"",methods:[],displayName:"FromDiviMarketPlace"};var d,l,u;a.parameters={...a.parameters,docs:{...(d=a.parameters)==null?void 0:d.docs,source:{originalSource:`() => {
+  const [licenseData, setLicenseData] = useState({
+    key: "",
+    licence_status: "",
+    licence_expire: "",
+    last_check: null
+  });
+  const LicensePageProps = {
+    isFromDiviMarket: false,
+    activateLicenseApi: "https://dummyjson.com/c/0a30-b68a-42b3-88c0",
+    forceLicenseCheckApi: "https://dummyjson.com/c/f33c-6d65-412e-8244",
+    deactivateLicenseApi: "https://dummyjson.com/c/03c3-39d0-45ae-9613",
+    licenseData: licenseData,
+    setLicenseData: (data: any) => setLicenseData(data),
+    findLicenseKeyLink: "https://www.peeayecreative.com/account/licence-keys/divi-assistant/",
+    docsLink: "https://www.peeayecreative.com/docs/divi-assistant/troubleshooting-faqs/"
+  };
+  return <LicensePage {...LicensePageProps} />;
+}`,...(u=(l=a.parameters)==null?void 0:l.docs)==null?void 0:u.source}}};var L,v,y;i.parameters={...i.parameters,docs:{...(L=i.parameters)==null?void 0:L.docs,source:{originalSource:`() => {
+  const [licenseData, setLicenseData] = useState({
+    key: "",
+    licence_status: "",
+    licence_expire: "",
+    last_check: null
+  });
+  const LicensePageProps = {
+    isFromDiviMarket: false,
+    activateLicenseApi: "https://dummyjson.com/c/21ea-fa9c-4fde-b725",
+    forceLicenseCheckApi: "https://dummyjson.com/c/f33c-6d65-412e-8244",
+    deactivateLicenseApi: "https://dummyjson.com/c/03c3-39d0-45ae-9613",
+    licenseData: licenseData,
+    setLicenseData: (data: any) => setLicenseData(data),
+    findLicenseKeyLink: " https://www.peeayecreative.com/account/licence-keys/divi-assistant/",
+    docsLink: "https://www.peeayecreative.com/docs/divi-assistant/troubleshooting-faqs/"
+  };
+  return <LicensePage {...LicensePageProps} />;
+}`,...(y=(v=i.parameters)==null?void 0:v.docs)==null?void 0:y.source}}};var h,k,f;n.parameters={...n.parameters,docs:{...(h=n.parameters)==null?void 0:h.docs,source:{originalSource:`() => {
+  const [licenseData, setLicenseData] = useState({
+    key: "PADTMAU-c26a58cc-e3b-2d756-b497591e",
+    licence_status: "active",
+    licence_expire: "2023-12-31",
+    last_check: Math.floor(Date.now() / 1000)
+  });
+  const LicensePageProps = {
+    isFromDiviMarket: false,
+    activateLicenseApi: "https://dummyjson.com/c/0a30-b68a-42b3-88c0",
+    forceLicenseCheckApi: "https://dummyjson.com/c/f33c-6d65-412e-8244-ss",
+    deactivateLicenseApi: "https://dummyjson.com/c/03c3-39d0-45ae-9613",
+    licenseData: licenseData,
+    setLicenseData: (data: any) => setLicenseData(data),
+    findLicenseKeyLink: " https://www.peeayecreative.com/account/licence-keys/divi-assistant/",
+    docsLink: "https://www.peeayecreative.com/docs/divi-assistant/troubleshooting-faqs/"
+  };
+  return <LicensePage {...LicensePageProps} />;
+}`,...(f=(k=n.parameters)==null?void 0:k.docs)==null?void 0:f.source}}};var D,w,g;o.parameters={...o.parameters,docs:{...(D=o.parameters)==null?void 0:D.docs,source:{originalSource:`() => {
+  return <LicensePage isFromDiviMarket={true} />;
+}`,...(g=(w=o.parameters)==null?void 0:w.docs)==null?void 0:g.source}}};const V=["NotActivated","InvalidKey","Activated","FromDiviMarketPlace"];export{n as Activated,o as FromDiviMarketPlace,i as InvalidKey,a as NotActivated,V as __namedExportsOrder,Q as default};